pyproject.toml
[project]
name = "tg-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mistralai (>=1.6.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "aiogram (>=3.19.0,<4.0.0)",
    "asyncio (>=3.4.3,<4.0.0)",
    "logging (>=0.4.9.6,<0.5.0.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)"
]

[tool.poetry]
packages = [{include = "tg_bot", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
